requirements.txt
groq>=0.4.0
ollama>=0.1.0
mistralai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0

